+++ src/components/photoCardList/components/Info.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import * as S from './Info.style';
 import { TypoGraphy } from 'src/components';
 import { customColor } from 'src/constants';
 import { FaHeart } from 'react-icons/fa';
+import styled from 'styled-components';
 
 type Props = {
   Area: string;
@@ -11,35 +11,60 @@ type Props = {
   Writer: string;
   Date: string;
 };
-export function Info({ Area, Like, Text, Writer, Date }: Props) {
+export function Info(props: Props) {
+  const { Area, Like, Text, Writer, Date } = props;
   return (
     <>
-      <S.ItemWrapper>
+      <ItemWrapper>
         <TypoGraphy fontWeight="bold" color={customColor.darkBrown} type="h3">
           {Area}
         </TypoGraphy>
-        <S.LikeWrapper>
-          <S.HeartWrapper>
+        <LikeWrapper>
+          <HeartWrapper>
             <FaHeart color="red" />
-          </S.HeartWrapper>
+          </HeartWrapper>
           <TypoGraphy fontWeight="bold" type="body2">
             {Like}
           </TypoGraphy>
-        </S.LikeWrapper>
-      </S.ItemWrapper>
-      <S.ContentWrapper>
+        </LikeWrapper>
+      </ItemWrapper>
+      <ContentWrapper>
         <TypoGraphy fontWeight="bold" fontHeight="1.2" type="body1">
           {Text}
         </TypoGraphy>
-      </S.ContentWrapper>
-      <S.ItemWrapper>
+      </ContentWrapper>
+      <ItemWrapper>
         <TypoGraphy fontWeight="bold" color={customColor.gray} type="body3">
           {Writer}
         </TypoGraphy>
         <TypoGraphy fontWeight="bold" color={customColor.gray} type="body3">
           {Date}
         </TypoGraphy>
-      </S.ItemWrapper>
+      </ItemWrapper>
     </>
   );
 }
+
+const ItemWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 10px 0;
+`;
+const LikeWrapper = styled.div`
+  display: flex;
+`;
+
+const ContentWrapper = styled.div`
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  text-overflow: ellipsis;
+  max-height: 2.4em;
+  overflow: hidden;
+  min-height: 40px;
+`;
+
+const HeartWrapper = styled.div`
+  margin-right: 2px;
+`;
