+++ src/components/photoCardList/components/PhotoCard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import * as S from './PhotoCard.style';
-import { Info, Tag } from '../index';
+import { Info, Tag } from './index';
 import { PhotoCardInfoType } from 'src/dummy/photoCardInfo';
+import styled from "styled-components";
 type Props = {
   photoCardInfo: PhotoCardInfoType;
 };
@@ -9,10 +9,22 @@ export const PhotoCard = ({ photoCardInfo }: Props) => {
   const { ImageUrl, Area, Like, Text, Writer, Date, Tags }: PhotoCardInfoType =
     photoCardInfo;
   return (
-    <S.Container>
-      <S.ImageWrapper ImageUrl={ImageUrl} />
+    <Container>
+      <ImageWrapper ImageUrl={ImageUrl} />
       <Info Area={Area} Like={Like} Text={Text} Writer={Writer} Date={Date} />
       <Tag Tags={Tags} />
-    </S.Container>
+    </Container>
   );
 };
+
+
+const Container = styled.div`
+  width: 260px;
+`;
+type ImageProps = {
+  ImageUrl?: string;
+};
+const ImageWrapper = styled.div<ImageProps>`
+  height: 180px;
+  background: ${({ ImageUrl }) => `url(${ImageUrl}) no-repeat`};
+`;
