--- src/components/photoCardList/components/PhotoCard/PhotoCard.style.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  width: 260px;
-`;
-type ImageProps = {
-  ImageUrl?: string;
-};
-export const ImageWrapper = styled.div<ImageProps>`
-  height: 180px;
-  background: ${({ImageUrl}) => `url(${ImageUrl}) no-repeat`};
-`;
